Fix for a bug when nvmf nexus child could not be removed
The problem was that we used a wrong name for destroying nvmf bdev.
To prevent similar bugs from appearing in the future the nexus tests
were improved to test all types of replica backends. The test code
related to nexus written in JS has been cleaned up & overhauled.

New spdk binary has been introduced for cases when just SPDK
functionality is desired without the overhead of mayastor. It is
used for creating nvmf target in the test suite (nvmf initiator
cannot be looped back to a target in the same process in SPDK).
