+++ mayastor-test/package.json
@@ -21,9 +21,6 @@
     "mocha": "^5.2.0",
     "wtfnode": "^0.8.0"
   },
-  "scripts": {
-    "test": "mocha test_cli.js; mocha test_grpc.js"
-  },
   "author": "Jan Kryl <jan.kryl@mayadata.io>",
   "license": "BSD-2-Clause"
 }
