+++ mayastor/src/nvme_dev.rs
@@ -157,9 +157,7 @@ impl NvmfBdev {
     /// destroy an nvme controller and its namespaces, it is not possible to
     /// destroy a nvme_bdev directly
     pub fn destroy(self) -> Result<(), nexus::Error> {
-        let mut name = self.name;
-        let name = name.split_off(name.len() - 2);
-        let cname = CString::new(name).unwrap();
+        let cname = CString::new(self.name).unwrap();
         let res = unsafe { spdk_sys::spdk_bdev_nvme_delete(cname.as_ptr()) };
 
         match res {
